Renamed namespace to ROSNET (#42)
